# Dense slotted array