--- a/src/Paprika.Tests/Data/SlottedArrayTests.cs
+++ b/src/Paprika.Tests/Data/SlottedArrayTests.cs
@@ -1,6 +1,5 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
-using Paprika.Crypto;
 using Paprika.Data;
 
 namespace Paprika.Tests.Data;
@@ -33,14 +32,14 @@ public class SlottedArrayTests
     }
 
     [Test]
-    public void Enumerate_all()
+    public void Enumerate_all([Values(0,1)] int odd)
     {
         Span<byte> span = stackalloc byte[256];
         var map = new SlottedArray(span);
 
         var key0 = Span<byte>.Empty;
-        Span<byte> key1 = stackalloc byte[1] { 7 };
-        Span<byte> key2 = stackalloc byte[2] { 7, 13 };
+        var key1 = NibblePath.FromKey(stackalloc byte[1] { 7 }).SliceFrom(odd);
+        var key2 = NibblePath.FromKey(stackalloc byte[2] { 7, 13 }).SliceFrom(odd);
 
         map.SetAssert(key0, Data0);
         map.SetAssert(key1, Data1);
@@ -53,11 +52,11 @@ public class SlottedArrayTests
         e.Current.RawData.SequenceEqual(Data0).Should().BeTrue();
 
         e.MoveNext().Should().BeTrue();
-        e.Current.Key.RawSpan.SequenceEqual(key1).Should().BeTrue();
+        e.Current.Key.Equals(key1).Should().BeTrue();
         e.Current.RawData.SequenceEqual(Data1).Should().BeTrue();
 
         e.MoveNext().Should().BeTrue();
-        e.Current.Key.RawSpan.SequenceEqual(key2).Should().BeTrue();
+        e.Current.Key.Equals(key2).Should().BeTrue();
         e.Current.RawData.SequenceEqual(Data2).Should().BeTrue();
 
         e.MoveNext().Should().BeFalse();
@@ -147,54 +146,60 @@ public class SlottedArrayTests
             map.GetAssert(key, value);
         }
     }
-
-    [Test]
-    public void Hashing()
-    {
-        var hashes = new Dictionary<ushort, string>();
-
-        // empty
-        Unique(ReadOnlySpan<byte>.Empty);
-
-        // single byte
-        Unique(stackalloc byte[] { 1 });
-        Unique(stackalloc byte[] { 2 });
-        Unique(stackalloc byte[] { 16 });
-        Unique(stackalloc byte[] { 17 });
-
-
-        // two bytes
-        Unique(stackalloc byte[] { 0xA, 0xC });
-        Unique(stackalloc byte[] { 0xA, 0xB });
-        Unique(stackalloc byte[] { 0xB, 0xC });
-
-        // three bytes
-        Unique(stackalloc byte[] { 0xA, 0xD, 0xC });
-        Unique(stackalloc byte[] { 0xAA, 0xEE, 0xCC });
-        Unique(stackalloc byte[] { 0xAA, 0xFF, 0xCC });
-
-        // three bytes
-        Unique(stackalloc byte[] { 0xAA, 0xDD, 0xCC, 0x11 });
-        Unique(stackalloc byte[] { 0xAA, 0xEE, 0xCC, 0x11 });
-        Unique(stackalloc byte[] { 0xAA, 0xFF, 0xCC, 0x11 });
-
-        return;
-
-        void Unique(in ReadOnlySpan<byte> key)
-        {
-            var hash = SlottedArray.GetHash(NibblePath.FromKey(key));
-            var hex = key.ToHexString(true);
-
-            if (hashes.TryAdd(hash, hex) == false)
-            {
-                Assert.Fail($"The hash for {hex} is the same as for {hashes[hash]}");
-            }
-        }
-    }
+    //
+    // [Test]
+    // public void Hashing()
+    // {
+    //     var hashes = new Dictionary<ushort, string>();
+    //
+    //     // empty
+    //     Unique(ReadOnlySpan<byte>.Empty);
+    //
+    //     // single byte
+    //     Unique(stackalloc byte[] { 1 });
+    //     Unique(stackalloc byte[] { 2 });
+    //     Unique(stackalloc byte[] { 16 });
+    //     Unique(stackalloc byte[] { 17 });
+    //
+    //
+    //     // two bytes
+    //     Unique(stackalloc byte[] { 0xA, 0xC });
+    //     Unique(stackalloc byte[] { 0xA, 0xB });
+    //     Unique(stackalloc byte[] { 0xB, 0xC });
+    //
+    //     // three bytes
+    //     Unique(stackalloc byte[] { 0xA, 0xD, 0xC });
+    //     Unique(stackalloc byte[] { 0xAA, 0xEE, 0xCC });
+    //     Unique(stackalloc byte[] { 0xAA, 0xFF, 0xCC });
+    //
+    //     // three bytes
+    //     Unique(stackalloc byte[] { 0xAA, 0xDD, 0xCC, 0x11 });
+    //     Unique(stackalloc byte[] { 0xAA, 0xEE, 0xCC, 0x11 });
+    //     Unique(stackalloc byte[] { 0xAA, 0xFF, 0xCC, 0x11 });
+    //
+    //     return;
+    //
+    //     void Unique(in ReadOnlySpan<byte> key)
+    //     {
+    //         var hash = SlottedArray.PrepareKey(NibblePath.FromKey(key));
+    //         var hex = key.ToHexString(true);
+    //
+    //         if (hashes.TryAdd(hash, hex) == false)
+    //         {
+    //             Assert.Fail($"The hash for {hex} is the same as for {hashes[hash]}");
+    //         }
+    //     }
+    // }
 }
 
 file static class FixedMapTestExtensions
 {
+    public static void SetAssert(this SlottedArray map, in NibblePath key, ReadOnlySpan<byte> data,
+        string? because = null)
+    {
+        map.TrySet(key, data).Should().BeTrue(because ?? "TrySet should succeed");
+    }
+    
     public static void SetAssert(this SlottedArray map, in ReadOnlySpan<byte> key, ReadOnlySpan<byte> data,
         string? because = null)
     {