@Scooletz 2024-03-05
![image](https://github.com/NethermindEth/Paprika/assets/519707/90dc49e4-fdba-4c5f-809b-933c1e23043d)

A snippet from syncing. We'll see.